+++ packages/univo/package.json
@@ -14,9 +14,9 @@
 			"types": "./dist/index.d.ts",
 			"import": "./dist/index.mjs"
 		},
-		"./client": {
-			"types": "./dist/client.d.ts",
-			"import": "./dist/client.mjs"
+		"./transport": {
+			"types": "./dist/transport.d.ts",
+			"import": "./dist/transport.mjs"
 		},
 		"./realtime": {
 			"types": "./dist/realtime.d.ts",
